Refactor - move custom hook to hooks folder - add index.ts to export all components
