+++ src/components/Loader/index.tsx
@@ -1,5 +1,5 @@
 import styles from './style.module.css'
 
-export default function Loader() {
+export function Loader() {
   return <div className={styles.loader}></div>;
 }
