+++ src/components/index.ts
@@ -1,0 +1,12 @@
+export * from './Button';
+export * from './Input';
+export * from './Loader';
+export * from './Modal';
+export * from './Header';
+export * from './Footer';
+export * from './Nav';
+export * from './NavItem';
+export * from './ProfileImg';
+export * from './UserCard';
+export * from './DashboardAdmin';
+export * from './DashboardUser';
